Split player service into smaller modules
